% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_get_municipality_info.R
\name{api_get_municipality_info}
\alias{api_get_municipality_info}
\title{Get municipality info from the SamenMeten API}
\usage{
api_get_municipality_info(municipality, conn)
}
\arguments{
\item{municipality}{Name of the municipality}

\item{conn}{Database connection}
}
\value{
List with municipality info
the function retrieves municipality information from the SamenMeten
API based on the provided municipality name. It uses the
municipality code to fetch the data and stores it in the database.
}
\description{
Get municipality info from the SamenMeten API
}
