% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_get_project_info.R
\name{api_get_project_info}
\alias{api_get_project_info}
\title{gets project info from the API Samen Meten}
\usage{
api_get_project_info(project, conn)
}
\arguments{
\item{project}{string name of the project}

\item{conn}{db connection object}
}
\value{
list with c("sensor_data", "datastream_data")
}
\description{
gets project info from the API Samen Meten
}
