% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_knmi.R
\name{download_data_knmi}
\alias{download_data_knmi}
\title{Download the data from KNMI}
\usage{
download_data_knmi(x, station, conn)
}
\arguments{
\item{x}{vector with the times, example c("20260501", "20260502")}

\item{station}{name of the station (string)}

\item{conn}{database connection}
}
\value{
dataframe wirth the columns c("station", "value", "timestamp",
"parameter", "aggregation")
}
\description{
Download the data from KNMI
}
\examples{
knmi_data <- download_data_knmi_edr(c("20260501", "20260502"), "KNMI_260")
}
