% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_location_knmi.R
\name{download_location_knmi}
\alias{download_location_knmi}
\title{Download location of KNMI station}
\usage{
download_location_knmi(knmi_station, time_start, time_end)
}
\arguments{
\item{knmi_station}{string with number of station (e.g. "KNMI_260")}

\item{time_start}{string with start time (posixct)}

\item{time_end}{string with end time (posixct)}
}
\value{
dataframe(station (string), lat (numeric), lon(numeric))
}
\description{
Download location of KNMI station
}
