% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_location_knmi_edr.R
\name{download_location_knmi_edr}
\alias{download_location_knmi_edr}
\title{Download location of KNMI station (using edr api)}
\usage{
download_location_knmi_edr(knmi_station, time_start, time_end)
}
\arguments{
\item{knmi_station}{string with number of station (e.g. "KNMI_260")}

\item{time_start}{string with start time (posixct)}

\item{time_end}{string with end time (posixct)}
}
\value{
dataframe(station (string), lat (numeric), lon(numeric))
}
\description{
Download location of KNMI station (using edr api)
}
\examples{
download_location_knmi_edr("KNMI_260",
as_datetime("2022-01-01 00:00:00"),
as_datetime("2022-01-08 23:59:59") )
}
