% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_locations_knmi.R
\name{download_locations_knmi}
\alias{download_locations_knmi}
\title{Download location of KNMI station}
\usage{
download_locations_knmi(knmi_stations, time_start, time_end)
}
\arguments{
\item{time_start}{string with start time (posixct)}

\item{time_end}{string with end time (posixct)}

\item{knmi_station}{string with number of station (e.g. "KNMI_260")}
}
\value{
dataframe(station (string), lat (numeric), lon(numeric))
}
\description{
Download location of KNMI station
}
\examples{
download_location_knmi("KNMI_260",
as_datetime("2022-01-01 00:00:00"),
as_datetime("2022-01-08 23:59:59") )
}
