% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_database_.R
\name{get_database_dirname}
\alias{get_database_dirname}
\title{Get database dirname
function to determine database dirname.}
\usage{
get_database_dirname()
}
\value{
This returns either the
database path as set in ANALYSETOGETHER_DATAFOLDER environment
variable, or returns default path (./data) if environment
variable is not present
}
\description{
Get database dirname
function to determine database dirname.
}
