% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_database_.R
\name{get_database_path}
\alias{get_database_path}
\title{Get database path
see function get_database_dirname. This function returns the
full path to the database}
\usage{
get_database_path(db = "database.db")
}
\arguments{
\item{db}{name of database file (sqlite database file)}
}
\value{
database path
}
\description{
Get database path
see function get_database_dirname. This function returns the
full path to the database
}
