% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_doc.R
\name{list_doc}
\alias{list_doc}
\title{List all documents of specific type}
\usage{
list_doc(type, conn)
}
\arguments{
\item{type}{string with type of document to list}

\item{conn}{database connection}
}
\value{
list of references with that type
}
\description{
List all documents of specific type
}
