% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_db.R
\name{reset_db}
\alias{reset_db}
\title{Reset db
removes the current database and creates a new one with the same}
\usage{
reset_db(fname)
}
\arguments{
\item{fname}{name of the database file (usually a temporary file)}
}
\value{
true if database was reset, false if not
}
\description{
Reset db
removes the current database and creates a new one with the same
}
