% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_to_days.R
\name{round_to_days}
\alias{round_to_days}
\title{Round to days
the samen meten API requires time ranges in full days. This
function rounds any time to the start of the day of time_start
until the end of the day of time_end}
\usage{
round_to_days(time_start, time_end)
}
\arguments{
\item{time_start}{string with start time (POSIXct)}

\item{time_end}{string with end time (POSIXct)}
}
\value{
c("time_start","time_end") with values rounded by day
}
\description{
Round to days
the samen meten API requires time ranges in full days. This
function rounds any time to the start of the day of time_start
until the end of the day of time_end
}
