% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_loglevel.R
\name{set_loglevel}
\alias{set_loglevel}
\title{Set loglevel
function to set the loglevel in the Analyse together tool,
using logger}
\usage{
set_loglevel(level = "INFO")
}
\arguments{
\item{level}{default "info" (options used: DEBUG, TRACE, INFO)}
}
\value{
no return value
}
\description{
Set loglevel
function to set the loglevel in the Analyse together tool,
using logger
}
