% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_exists.R
\name{station_exists}
\alias{station_exists}
\title{Check if station in database exists
checks if 'station' already exists in the location table}
\usage{
station_exists(station, conn)
}
\arguments{
\item{station}{string name of station}

\item{conn}{database connection}
}
\value{
TRUE of FALSE
}
\description{
Check if station in database exists
checks if 'station' already exists in the location table
}
