% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_knmi_edr.R
\name{wrap_GetKNMIAPIEDR}
\alias{wrap_GetKNMIAPIEDR}
\title{Wrapper around GetKNMIAPIEDR
This function is a wrapper around GetKNMIAPIEDR and includes
the other input variables, like token, parameter and data_result.}
\usage{
wrap_GetKNMIAPIEDR(station, ts_api, te_api)
}
\arguments{
\item{station}{station name}

\item{ts_api}{start date}

\item{te_api}{end date}
}
\value{
dataframe with the KNMI data in columns c(values (numeric),
date_time (posixct), id_nr(character), parameter_name (character),
result_type (character), lat (numeric), lon(numeric). If not succesful
call then NULL is returned
}
\description{
see samanapir::GetKNMIAPIEDR for the full info
}
