% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_get_municipality_info.R
\name{wrap_GetSamenMetenAPIinfoMuni}
\alias{wrap_GetSamenMetenAPIinfoMuni}
\title{Wrapper around GetSamenMetenAPIinfoMuni
This function is a wrapper around samanapir::GetSamenMetenAPIinfoMuni2
mostly for testing}
\usage{
wrap_GetSamenMetenAPIinfoMuni(gemid)
}
\arguments{
\item{gemid}{municipality code}
}
\value{
list
}
\description{
See samanapir::GetSamenMetenAPIinfoMuni2, for more info
}
