% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_lml.R
\name{download_data_lml}
\alias{download_data_lml}
\title{Download the data from Luchtmeetnet}
\usage{
download_data_lml(x, station)
}
\arguments{
\item{x}{vector with the times, example c("20260501", "20260502")}

\item{station}{name of the station (string)}

\item{conn}{database connection}
}
\value{
dataframe wirth the columns ("station", "value", "timestamp",
"parameter", "timestamp_measured_start", "timestamp_measured_end", "aggregation"))
}
\description{
Download the data from Luchtmeetnet
}
\examples{
lml_data <- download_data_lml(c("20260501", "20260502"), "NL49680")
}
