% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_sensor_meta.R
\name{download_sensor_meta}
\alias{download_sensor_meta}
\title{Download meta data sensor Samen Meten
This function downloads a set of sensors belonging to either a
project or municipality,  download the station meta data, and
downloads the measurements for the requested time range
The type arguments determine if data is requested for a
municpality or a project.}
\usage{
download_sensor_meta(name, type, conn)
}
\arguments{
\item{name}{name of project or municipality}

\item{type}{either 'project' or 'municipality'}

\item{conn}{db connection object}
}
\value{
df with the columns c("station", "parameter",
"timestamp", "value", "aggregation")
}
\description{
Download meta data sensor Samen Meten
This function downloads a set of sensors belonging to either a
project or municipality,  download the station meta data, and
downloads the measurements for the requested time range
The type arguments determine if data is requested for a
municpality or a project.
}
\examples{
 download_sensor_meta("Amersfoort", type = "municipality", conn = dbconn)
}
