% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_knmi.R
\name{wrap_GetKNMIAPI}
\alias{wrap_GetKNMIAPI}
\title{Wrapper around GetKNMIAPI
This function is a wrapper around GetKNMIAPI and includes
the other input variables, like token, parameter and data_result.}
\usage{
wrap_GetKNMIAPI(station, ts_api, te_api)
}
\arguments{
\item{station}{station name}

\item{ts_api}{start date}

\item{te_api}{end date}
}
\value{
dataframe
}
\description{
see samanapir::GetKNMIAPI for the full info
}
