% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_lml.R
\name{wrap_GetLMLstatdataAPI}
\alias{wrap_GetLMLstatdataAPI}
\title{Wrapper around GetLMLstatdataAPI2
This function is a wrapper around GetLMLstatdataAPI2 and includes
the other input variables, like token, parameter and data_result.}
\usage{
wrap_GetLMLstatdataAPI(station, ts_api, te_api)
}
\arguments{
\item{station}{station name}

\item{ts_api}{start date}

\item{te_api}{end date}
}
\value{
dataframe
}
\description{
see samanapir::GetLMLstatdataAPI2 for the full info
}
