% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_get_project_info.R
\name{wrap_GetSamenMetenAPIinfoProject}
\alias{wrap_GetSamenMetenAPIinfoProject}
\title{Wrapper around GetSamenMetenAPIinfoProject
This function is a wrapper around samanapir::GetSamenMetenAPIinfoProject
mostly for testing}
\usage{
wrap_GetSamenMetenAPIinfoProject(project)
}
\arguments{
\item{project}{project name}
}
\value{
list
}
\description{
See samanapir::GetSamenMetenAPIinfoProject, for more info
}
